FastAPI/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
mysql-connector-python==9.0.0
google-generativeai==0.7.2
